feat: 피드 삭제 기능 구현 (#7)

+++ src/main/java/com/example/feedserver/feed/service/SocialFeedService.java
@@ -1,10 +1,11 @@
 package com.example.feedserver.feed.service;
 
+import com.example.feedserver.exception.BadRequestException;
 import com.example.feedserver.feed.dto.CreateFeedRequest;
 import com.example.feedserver.feed.dto.FeedInfo;
 import com.example.feedserver.feed.dto.FeedResponse;
-import com.example.feedserver.feed.repository.SocialFeedRepository;
 import com.example.feedserver.feed.entity.SocialFeed;
+import com.example.feedserver.feed.repository.SocialFeedRepository;
 import com.example.feedserver.userinfo.UserInfo;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -24,7 +25,7 @@ public class SocialFeedService {
 
     private final SocialFeedRepository feedRepository;
     private final KafkaTemplate<String, String> kafkaTemplate;
-    private ObjectMapper objectMapper = new ObjectMapper();
+    private final ObjectMapper objectMapper;
 
     @Value("${sns.user-server}")
     private String userServerUrl;
@@ -44,24 +45,50 @@ public class SocialFeedService {
         return convertToSocialFeedResponse(socialFeed);
     }
 
-    public void deleteFeed(int feedId) {
-        feedRepository.deleteById(feedId);
-    }
-
     @Transactional
-    public FeedResponse createFeed(CreateFeedRequest feedRequest) {
+    public FeedResponse createFeed(CreateFeedRequest feedRequest, int userId) {
+        if (feedRequest.getUploaderId() != userId) {
+            throw new BadRequestException();
+        }
         SocialFeed feed = feedRepository.save(convertToSocialFeed(feedRequest));
 
         UserInfo uploader = getUserInfo(feed.getUploaderId());
         FeedInfo feedInfo = new FeedInfo(feed, uploader.getUsername());
 
+        sendCreateMessageToKafka(feedInfo);
+
+        return convertToSocialFeedResponse(feed);
+    }
+
+    public void deleteFeed(int feedId, int userId) {
+        SocialFeed feed = feedRepository.findById(feedId)
+                .orElseThrow(() -> new BadRequestException());
+
+        if (feed.getUploaderId() != userId) {
+            throw new BadRequestException();
+        }
+
+        UserInfo userInfo = getUserInfo(userId);
+        FeedInfo feedInfo = new FeedInfo(feed, userInfo.getUsername());
+
+        feedRepository.deleteById(feedId);
+        sendDeleteMessageToKafka(feedInfo);
+    }
+
+    private void sendCreateMessageToKafka(FeedInfo feedInfo) {
         try {
             kafkaTemplate.send("feed.created", objectMapper.writeValueAsString(feedInfo));
         } catch (JsonProcessingException e) {
             throw new RuntimeException(e);
         }
+    }
 
-        return convertToSocialFeedResponse(feed);
+    private void sendDeleteMessageToKafka(FeedInfo feedInfo) {
+        try {
+            kafkaTemplate.send("feed.deleted", objectMapper.writeValueAsString(feedInfo));
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     public void refreshAllFeeds() {
@@ -70,11 +97,17 @@ public class SocialFeedService {
             UserInfo uploader = getUserInfo(feed.getUploaderId());
             FeedInfo feedInfo = new FeedInfo(feed, uploader.getUsername());
 
-            try {
-                kafkaTemplate.send("feed.created", objectMapper.writeValueAsString(feedInfo));
-            } catch (JsonProcessingException e) {
-                throw new RuntimeException(e);
-            }
+            sendCreateMessageToKafka(feedInfo);
+        }
+    }
+
+    public void deleteAllFeeds() {
+        List<SocialFeed> feeds = getAllFeeds();
+        for (SocialFeed feed: feeds) {
+            UserInfo uploader = getUserInfo(feed.getUploaderId());
+            FeedInfo feedInfo = new FeedInfo(feed, uploader.getUsername());
+
+            sendDeleteMessageToKafka(feedInfo);
         }
     }
